Moving tab & app -> tui/
