+++ src/tui/mod.rs
@@ -1,0 +1,4 @@
+pub mod app;
+mod tab;
+
+pub use app::App;
